move genpool image from donate page to main page
