+++ app/components/Features/DonateForm/index.js
@@ -15,7 +15,6 @@ import CardGiftcard from '@material-ui/icons/CardGiftcard';
 import Tool from 'components/Tool/Tool';
 import ToolSection from 'components/Tool/ToolSection';
 import ToolBody from 'components/Tool/ToolBody';
-import GridItem from "components/Grid/GridItem";
 
 import Donate from 'components/Information/Donate';
 
@@ -23,18 +22,6 @@ import FormObject from './FormObject';
 
 import messages from './messages';
 import commonMessages from '../../messages';
-import genpoolWeb from '../../../assets/img/genpool.png';
-
-const bannerBackground = {
-  boxShadow: '0 1px 4px 0 rgba(0, 0, 0, 0.14)',
-  width: '100%',
-};
-
-const bannerImage = {
-  margin: '0 auto',
-  borderRadius: 6,
-  width: '100%',
-};
 
 const makeTransaction = (values, networkIdentity) => {
   const transaction = [
@@ -58,13 +45,6 @@ const DonateForm = props => {
   const { intl } = props;
   return (
     <Tool>
-      <GridItem xs={12} sm={12} md={12} lg={12}>
-        <div style={bannerBackground}>
-          <a target="_blank" href="https://genpool.io/">
-            <img style={bannerImage} src={genpoolWeb} alt="eosbot-banner" />
-          </a>
-        </div>
-      </GridItem>
       <ToolSection lg={12}>
         <ToolBody color="warning"
                   icon={CardGiftcard}
